changed remove
can't really do the column/ row thingy here because the max number of columns in a row can be changed. Can thus be problematic for user. FixedLengthTable and StatisticalTable (will be developed in the future) won't have any features changing the structure of the table representation and can use the column/row approch to find where an element is. But this class is not suited for it!